lib8 (keyboard, text screen)

+++ nekot1/prelude.c
@@ -52,7 +52,7 @@ void* memcpy(void* dest, const void* src, size_t count) {
     return dest;
 }
 
-size_t strlen(const char* s) {
+int strlen(const char* s) {
     const char* p = s;
     while (*p) ++p;
     return p-s;
